==> STEP1_pdf_to_text_non_retracted.ipynb <==
Connect to drive

from google.colab import drive
drive.mount('/content/drive')

Set directory of the folder where the downloads are placed in

cd '/content/drive/MyDrive/Applied Data Science/Thesis/Code/Downloads (paper pdfs)/Downloads_Non_Retracted_11_Journals'

import os
files = os.listdir('.')

!pip install pymupdf

import pandas as pd
import fitz  # this is pymupdf
df_non_retracted = pd.DataFrame(columns=['DOI', 'Content', 'Retracted']) #make empty dataframe
for study in files: #iterate through the files
  try:
    with fitz.open(study) as doc: #open the document
        text = "" #start with empty string
        for page in doc: #iterate through the pages of the file
            text += page.get_text() #add the found text to the empty string
        df_non_retracted = df_non_retracted.append({'DOI': study, 'Content': text, 'Retracted': 0}, ignore_index=True) #add DOI, text and 0 for non-retracted paper to the dataframe
  except: #if anything above fails, pass this file
    pass

df_non_retracted

OPTIONAL:
Subsetting dataset:
Dataset now gets equally amounts as the retracted database has

a= df_non_retracted.iloc[0:340] #340 from the first journal
b = df_non_retracted.iloc[1235:1428] # 193 from the second journal
print(len(a))
print(len(b))

df_non_retracted_2_papers = pd.concat([a,b])

df_non_retracted_2_papers

Save dataframe to a csv file

df_non_retracted.to_csv('/content/drive/MyDrive/Applied Data Science/Thesis/Code/Data (CSV)/non_retracted_11_journals.csv', index=False, encoding = 'utf-8-sig')
