==> STEP1_pdf_to_text_retracted.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Applied Data Science/Thesis/Code/Downloads (paper pdfs)/Downloads_Retracted_11_Journals'

import os
files = os.listdir('.')

!pip install pymupdf

import pandas as pd
import fitz  # this is pymupdf
df_retracted = pd.DataFrame(columns=['DOI', 'Content', 'Retracted'])
for study in files:
  try:
    with fitz.open(study) as doc:
        text = ""
        for page in doc:
            text += page.get_text()
        df_retracted = df_retracted.append({'DOI': study, 'Content': text, 'Retracted': 1}, ignore_index=True)
  except:
    pass

df_retracted

df_retracted.to_csv('/content/drive/MyDrive/Applied Data Science/Thesis/Code/Data (CSV)/retracted_11_journals.csv', index=False, encoding = 'utf-8-sig')
