==> STEP2_Creating Balanced Dataset.ipynb <==
## Meta Concat Retracted

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Applied Data Science/Thesis/Code/Data (CSV)'

import pandas as pd

df_11_journals = pd.read_csv('retracted_11_journals.csv')
df_meta_11_journals = pd.read_csv('meta_retracted_11_journals.csv')

df_11_journals['ID'] = df_11_journals.DOI.str.replace('.pdf', '', regex = True)

r_result = pd.merge(df_11_journals, df_meta_11_journals, on="ID")

r_result.Journal.value_counts()

## Meta Concat Non-Retracted

import pandas as pd

df_11_journals_non_retracted = pd.read_csv('non_retracted_11_journals.csv')
df_meta_11_journals_non_retracted = pd.read_csv('meta_non_retracted_11_journals_extra.csv')

df_11_journals_non_retracted['ID'] = df_11_journals_non_retracted.DOI.str.replace('.pdf', '', regex = True)

nr_result = pd.merge(df_11_journals_non_retracted, df_meta_11_journals_non_retracted, on="ID")

nr_result['Source Title'].value_counts()

## Balanced dataset

nr_df_fas  = nr_result[nr_result['Source Title'].str.contains("JOURNAL OF FUNDAMENTAL AND APPLIED SCIENCES", na=False)]
nr_df_plos = nr_result[nr_result['Source Title'].str.contains("PLOS ONE", na=False)]
nr_df_ajg  = nr_result[nr_result['Source Title'].str.contains("ARABIAN JOURNAL OF GEOSCIENCES", na=False)]
nr_df_cb  = nr_result[nr_result['Source Title'].str.contains("JOURNAL OF CELLULAR BIOCHEMISTRY", na=False)]
nr_df_ac  = nr_result[nr_result['Source Title'].str.contains("ACTA CRYSTALLOGRAPHICA SECTION E-CRYSTALLOGRAPHIC COMMUNICATIONS", na=False)]
nr_df_ra  = nr_result[nr_result['Source Title'].str.contains("RSC ADVANCES", na=False)]
nr_df_jca  = nr_result[nr_result['Source Title'].str.contains("JOURNAL OF CLINICAL ANESTHESIA", na=False)]
nr_df_ot   = nr_result[nr_result['Source Title'].str.contains("ONCOTARGETS AND THERAPY", na=False)]
nr_df_ijee = nr_result[nr_result['Source Title'].str.contains("INTERNATIONAL JOURNAL OF ELECTRICAL ENGINEERING EDUCATION", na=False)]



nr_balanced = pd.concat([nr_df_fas ,
nr_df_plos,
nr_df_ajg ,
nr_df_cb  ,
nr_df_ac  ,
nr_df_ra  ,
nr_df_jca ,
nr_df_ot  ,
nr_df_ijee])

r_df_fas = r_result[r_result.Journal.str.contains("Journal of Fundamental and Applied Sciences", na=False)]
r_df_plos = r_result[r_result.Journal.str.contains("PLoS One", na=False)]
r_df_ajg = r_result[r_result.Journal.str.contains("Arabian Journal of Geosciences", na=False)]
r_df_cb = r_result[r_result.Journal.str.contains("Journal of Cellular Biochemistry", na=False)]
r_df_ac = r_result[r_result.Journal.str.contains("Acta Crystallographica. Section E, Crystallographic Communications", na=False)]
r_df_ra = r_result[r_result.Journal.str.contains("RSC Advances", na=False)]
r_df_jca = r_result[r_result.Journal.str.contains("Journal of Clinical Anesthesia", na=False)]
r_df_ot = r_result[r_result.Journal.str.contains("OncoTargets and Therapy", na=False)]
r_df_ijee = r_result[r_result.Journal.str.contains("International Journal of Electrical Engineering & Education", na=False)]

r_balanced = pd.concat([r_df_fas, 
r_df_plos,
r_df_ajg,
r_df_cb,
r_df_ac,
r_df_ra,
r_df_jca,
r_df_ot,
r_df_ijee])

r_balanced.Journal.value_counts()

nr_balanced['Source Title'].value_counts()

nr_balanced.to_csv('balanced_non_retracted_11_journals.csv')
r_balanced.to_csv('balanced_retracted_11_journals.csv')
