==> STEP3_preprocessing part 1.ipynb <==
Connect to drive

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Applied Data Science/Thesis/Code'

Import Libraries

import pandas as pd
import numpy as np
import re

Load csv files of retracted and non-retracted articles

non_retracted = pd.read_csv('Data (CSV)/balanced_non_retracted_11_journals.csv', encoding="utf-8-sig")

retracted = pd.read_csv('Data (CSV)/balanced_retracted_11_journals.csv', encoding="utf-8-sig")

Some words have been cut off with a - and new line. To get the original words, the -\n is removed

non_retracted = non_retracted.replace({'-\n': ''}, regex = True)

retracted = retracted.replace({'-\n': ''}, regex = True)

Remove \n for both retracted and non-retracted papers

non_retracted = non_retracted.replace({'\n': ' '}, regex = True)

retracted = retracted.replace({'\n': ' '}, regex = True)

Concat to one dataframe

papers = pd.concat([non_retracted, retracted], ignore_index=True)

Remove rows with missing values

papers['Content'].dropna(inplace=True)

Original scraped content is kept as OG_content.

papers['All content'] = papers['Content']

Try to remove retraction notice

len(papers[papers['All content'].str.contains('Notice of Retraction')==True])

papers.Content = papers.Content.str.replace(r'^Notice of Retraction.+ieee\.org\.', '', regex = True)

Remove papers from non-retracted that are actually retracted

len(papers)

non_retracted = papers[papers.Retracted == 0]
doi_toremove = non_retracted[non_retracted["Content"].str.contains('Retraction|retraction|Retracted|retracted|retract|Retract|withdrawn|Withdraw')==True].DOI_x
papers = papers[~papers.DOI_x.isin(doi_toremove)]

len(papers)

Splitting the content into 4 parts: 1) Abstract, 2) main content (Introduction-Results), 3) Discussion/Conclusion and 4) References. 

papers['References'] = papers['Content'].str.split(r'References|R E F E R E N C E S|REFERENCES|r e f e r e n c e s|Reference List', 2, expand = True)[1]
papers['Content'] = papers['Content'].str.split(r'References|R E F E R E N C E S|REFERENCES|r e f e r e n c e s|Reference List', 2, expand = True)[0]

papers['Discussion / Conclusion'] = papers['Content'].str.split(r'Discussion|D I S C U S S I O N|DISCUSSION|d i s c u s s i o n|Conclusion|C O N C L U S I O N|CONCLUSION|c o n c l u s i o n', 2, expand = True)[1]
papers['Content'] = papers['Content'].str.split(r'Discussion|D I S C U S S I O N|DISCUSSION|d i s c u s s i o n|Conclusion|C O N C L U S I O N|CONCLUSION|c o n c l u s i o n', 2, expand = True)[0]

papers['Main content'] = papers['Content'].str.split(r'Introduction|I N T R O D U C T I O N|INTRODUCTION|i n t r o d u c t i o n', 2, expand = True)[1]
papers['Title + Abstract'] = papers['Content'].str.split(r'Introduction|I N T R O D U C T I O N|INTRODUCTION|i n t r o d u c t i o n', 2, expand = True)[0]
papers.drop('Content', axis=1, inplace=True)

papers

Check for null values (NaN/None) in References: if references are missing, it is an indicator that not the original article was scraped but some text relating to the article. Thus, we decide to exclude those rows with missing values for references.

print("Total papers:", len(papers))
print("Missing main content:", papers['Main content'].isnull().sum())
print("Missing discussion/conclusion:", papers['Discussion / Conclusion'].isnull().sum())
print("Missing references:", papers['References'].isnull().sum())

no_references = papers[papers['References'].isnull()]
papers = papers[~papers.ID.isin(no_references.ID)]

print("Total papers:", len(papers))
print("Missing main content:", papers['Main content'].isnull().sum())
print("Missing discussion/conclusion:", papers['Discussion / Conclusion'].isnull().sum())
print("Missing references:", papers['References'].isnull().sum())

no_introduction = papers[papers['Main content'].isnull()]
papers = papers[~papers.ID.isin(no_introduction.ID)]

print("Total papers:", len(papers))
print("Missing main content:", papers['Main content'].isnull().sum())
print("Missing discussion/conclusion:", papers['Discussion / Conclusion'].isnull().sum())
print("Missing references:", papers['References'].isnull().sum())

no_discussion = papers[papers['Discussion / Conclusion'].isnull()]
papers = papers[~papers.ID.isin(no_discussion.ID)]

print("Total papers:", len(papers))
print("Missing main content:", papers['Main content'].isnull().sum())
print("Missing discussion/conclusion:", papers['Discussion / Conclusion'].isnull().sum())
print("Missing references:", papers['References'].isnull().sum())

Remove duplicate entries

duplicates = []
count = 0
for i in range(len(papers['Main content'])):
  count += 1
  for j in range(len(papers['Main content'])):
    try:
      if papers['Main content'][i] == papers['Main content'][j] and i < j and str(papers['Main content'][i]) != 'None':
        duplicates.append(papers['ID'][i])
    except:
      pass

papers = papers[~papers.ID.isin(duplicates)]

print("Total papers after duplicate removal:", len(papers))

Investigate the final dataframe

papers['Journal_Name'] = papers['Journal'].fillna(papers['Source Title'])

papers['Journal_Name']  = papers['Journal_Name'].str.lower()

papers['Journal_Name'] = papers['Journal_Name'].str.replace('acta crystallographica section e-crystallographic communications', 'acta crystallographica. section e, crystallographic communications', regex = True)
papers['Journal_Name'] = papers['Journal_Name'].str.replace('international journal of electrical engineering & education', 'international journal of electrical engineering education', regex = True)

papers['Journal_Name'].value_counts()

papers.groupby(['Journal_Name', 'Retracted', ]).size()

Save the output to CSV

papers.to_csv('/content/drive/MyDrive/Applied Data Science/Thesis/Code/Data (CSV)/preprocessing_part_1.csv')
