==> STEP4_preprocessing part 2.ipynb <==
Connect to drive

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Applied Data Science/Thesis/Code'

Import Libraries

import pandas as pd
import numpy as np
import re

Load csv files of retracted and non-retracted articles

papers = pd.read_csv('Data (CSV)/preprocessing_part_1.csv', encoding="utf-8-sig")

## Preprocessing (V1) for all except Bert: Lowercasing, retracted_word removal, numbers removal, journal removal, author removal

#### Retracted word removal

papers['Title + Abstract'] = papers['Title + Abstract'].str.replace('Retraction|retraction|Retracted|retracted|retract|Retract|withdrawn|Withdraw', '', regex = True)
papers['Main content'] = papers['Main content'].str.replace('Retraction|retraction|Retracted|retracted|retract|Retract|withdrawn|Withdraw', '', regex = True)
papers['Discussion / Conclusion'] = papers['Discussion / Conclusion'].str.replace('Retraction|retraction|Retracted|retracted|retract|Retract|withdrawn|Withdraw', '', regex = True)
papers['References'] = papers['References'].str.replace('Retraction|retraction|Retracted|retracted|retract|Retract|withdrawn|Withdraw', '', regex = True)

#### Number removal

papers['Title + Abstract'] = papers['Title + Abstract'].str.replace(r'\d', '', regex = True)
papers['Main content'] = papers['Main content'].str.replace(r'\d', '', regex = True)
papers['Discussion / Conclusion'] = papers['Discussion / Conclusion'].str.replace(r'\d', '', regex = True)
papers['References'] = papers['References'].str.replace(r'\d', '', regex = True)

#### Author, Journal, Space, Stop word, Punctuation removal + lowercasing & lemmatization
Note: only abbreviations remain 

import spacy 
import re

nlp = spacy.load("en_core_web_sm")

Preprocessed Large

sections = ['Title + Abstract', 'Main content', 'Discussion / Conclusion', 'References']
for section in sections:
  texts = list(papers[section].values)
  processed_texts = [text for text in nlp.pipe(texts, disable=["ner", "parser"])]

  preprocessed_texts = [[word.lemma_.lower() for word in processed_text if not word.is_stop and not word.is_punct and not word.pos_ == "PROPN" and not word.is_space] 
                        for processed_text in processed_texts]

  newcolumn = section + ' ' + 'PP L'

  papers[newcolumn] = preprocessed_texts
  papers = papers.astype({newcolumn: str}, errors='raise') 
  papers[newcolumn] = papers[newcolumn].replace({',': ''}, regex = True)
  papers[newcolumn] = papers[newcolumn].replace({'\'': ''}, regex = True)
  papers[newcolumn] = papers[newcolumn].replace({'\[': ''}, regex = True)
  papers[newcolumn] = papers[newcolumn].replace({'\]': ''}, regex = True)

Preprocessed Small - only removal of proper nouns to remove all author names, journal titles and white spaces

sections = ['Title + Abstract', 'Main content', 'Discussion / Conclusion', 'References']
for section in sections:
  texts = list(papers[section].values)
  processed_texts = [text for text in nlp.pipe(texts, disable=["ner", "parser"])]

  preprocessed_texts = [[word for word in processed_text if not word.pos_ == "PROPN" and not word.is_space] 
                        for processed_text in processed_texts]

  newcolumn = section + ' ' + 'PP S'

  papers[newcolumn] = preprocessed_texts
  papers = papers.astype({newcolumn: str}, errors='raise') 
  papers[newcolumn] = papers[newcolumn].replace({',': ''}, regex = True)
  papers[newcolumn] = papers[newcolumn].replace({'\'': ''}, regex = True)
  papers[newcolumn] = papers[newcolumn].replace({'\[': ''}, regex = True)
  papers[newcolumn] = papers[newcolumn].replace({'\]': ''}, regex = True)

#### Creating common column for journal name

papers['Journal_Name'] = papers['Journal'].fillna(papers['Source Title'])

papers['Journal_Name']  = papers['Journal_Name'].str.lower()

papers['Journal_Name'] = papers['Journal_Name'].str.replace('acta crystallographica section e-crystallographic communications', 'acta crystallographica. section e, crystallographic communications', regex = True)
papers['Journal_Name'] = papers['Journal_Name'].str.replace('international journal of electrical engineering & education', 'international journal of electrical engineering education', regex = True)

papers['Journal_Name'].value_counts()

papers.groupby(['Journal_Name', 'Retracted', ]).size()

#### Balance data based on journal

#r_acta  = papers[(papers['Journal_Name'].str.contains("acta crystallographica. section e, crystallographic communications", na=False)) & (papers['Retracted']== 1)].iloc[:50]
#nr_acta  = papers[(papers['Journal_Name'].str.contains("acta crystallographica. section e, crystallographic communications", na=False)) & (papers['Retracted']== 0)].iloc[:50]

r_arabian  = papers[(papers['Journal_Name'].str.contains("arabian journal of geosciences", na=False)) & (papers['Retracted']== 1)].iloc[:47]
nr_arabian  = papers[(papers['Journal_Name'].str.contains("arabian journal of geosciences", na=False)) & (papers['Retracted']== 0)].iloc[:47]

r_engineering  = papers[(papers['Journal_Name'].str.contains("international journal of electrical engineering education", na=False)) & (papers['Retracted']== 1)].iloc[:32]
nr_engineering  = papers[(papers['Journal_Name'].str.contains("international journal of electrical engineering education", na=False)) & (papers['Retracted']== 0)].iloc[:32]

r_cellular_biochem  = papers[(papers['Journal_Name'].str.contains("journal of cellular biochemistry", na=False)) & (papers['Retracted']== 1)].iloc[:78]
nr_cellular_biochem  = papers[(papers['Journal_Name'].str.contains("journal of cellular biochemistry", na=False)) & (papers['Retracted']== 0)].iloc[:78]

#r_clin_anesthesia  = papers[(papers['Journal_Name'].str.contains("journal of clinical anesthesia", na=False)) & (papers['Retracted']== 1)].iloc[:80]
#nr_clin_anesthesia  = papers[(papers['Journal_Name'].str.contains("journal of clinical anesthesia", na=False)) & (papers['Retracted']== 0)].iloc[:80]

r_fund_app  = papers[(papers['Journal_Name'].str.contains("journal of fundamental and applied sciences", na=False)) & (papers['Retracted']== 1)].iloc[:12]
nr_fund_app  = papers[(papers['Journal_Name'].str.contains("journal of fundamental and applied sciences", na=False)) & (papers['Retracted']== 0)].iloc[:12]

r_onco  = papers[(papers['Journal_Name'].str.contains("oncotargets and therapy", na=False)) & (papers['Retracted']== 1)].iloc[:13]
nr_onco  = papers[(papers['Journal_Name'].str.contains("oncotargets and therapy", na=False)) & (papers['Retracted']== 0)].iloc[:13]

r_plos  = papers[(papers['Journal_Name'].str.contains("plos one", na=False)) & (papers['Retracted']== 1)].iloc[:60]
nr_plos  = papers[(papers['Journal_Name'].str.contains("plos one", na=False)) & (papers['Retracted']== 0)].iloc[:60]

r_rsc  = papers[(papers['Journal_Name'].str.contains("rsc advances", na=False)) & (papers['Retracted']== 1)].iloc[:72]
nr_rsc  = papers[(papers['Journal_Name'].str.contains("rsc advances", na=False)) & (papers['Retracted']== 0)].iloc[:72]


#### Creating subset (7/2) 

Note: so we can also see if classifiers are accurate for distinguishing on retraction for journals they are not trained on

two_journal_external_data_set = pd.concat([r_plos, nr_plos, r_rsc, nr_rsc])
five_journal_train_test_data_set = pd.concat([r_cellular_biochem, nr_cellular_biochem, r_engineering, nr_engineering, r_fund_app, nr_fund_app, r_onco, nr_onco, r_arabian, nr_arabian])

two_journal_external_data_set.groupby(['Journal_Name', 'Retracted', ]).size()

five_journal_train_test_data_set.groupby(['Journal_Name', 'Retracted', ]).size()

two_journal_external_data_set.to_csv('/content/drive/MyDrive/Applied Data Science/Thesis/Code/Data (CSV)/two_journal_external_data_set.csv')
five_journal_train_test_data_set.to_csv('/content/drive/MyDrive/Applied Data Science/Thesis/Code/Data (CSV)/five_journal_train_test_data_set.csv')
